git-rebase: disable save check when showing commits
The git-rebase-todo buffer is frequently in a modified state, but there
is no need to save it before showing a commit.  It will be taken care of
by with-editor-return when the user exits through with-editor-finish.

The unsaved buffer should almost always be the git-rebase-todo because
the initial 'git rebase' call will fail if there are unstaged changes.
The user could unwisely go and modify some other tracked buffers after
starting a rebase, and this could put things into a bad state [1], but
this is unrelated to whether the user happened to call a command to show
a commit.

Closes #2770.

[1] And, when the user has done this, I don't think there is really any
    good action to take to fix the situation.  It's bad whether or not
    the buffer is saved.  The user should just never start an
    interactive rebase and then change buffers of tracked files before
    triggering the rebase with with-editor-finish.