Proposal
Use an alist to store all properties of the worktree (without
hard-coding the list of supported properties).

Use this full worktree object in the definition of
`magit-insert-worktrees' -- using an additional alist
key (`magit-label') to store the propertized branch name / commit OID
so that it can be passed to the single-section-inserter function
embedded in the rest of the worktree object.

This also avoids the allocation of the separate `cols' list, FWIW.